Minor tweaks to event spec selectors
The events section recently changed from a `h2` to a `h4` tag for better
document structure, this reflects the change in the event specs.

Fast-forward month to November and next month to December (as there are no
longer any events in October). We will want a more dynamic way to set these
going forward; we could work it out by querying the API or stepping through the
months until we hit one with results.