Remove redundant home page specs
Most of these are testing internal/external links, for which we have tests
already for every link on the website.

We find that these tests often break because the content changes, and we
probably don't want to be notified if a link on the website changes (assuming
its still pointing at a valid page).