Disable event sign up tests
The event sign up tests require Train to Teach events to be present (as they
are the only event type that we accept online sign ups for).

Currently there are 0 TTT events live, so the sign ups are all failing.

Disable the tests until we have TTT events live again.