Add minimal integration tests
We see a lot of flakiness around the current tests and they also
test areas that are covered already by the tests within the apps
themselves.

Instead, we want to use Cypress to only do a minimal integration
test that proves the entire system is working as a whole.

This commit introduces a minimal sign up journey for mailing list
and get an adviser services, covering new candidates and the
matchback process for existing candidates.