Add Train to Teach event sign up tests
Add a minimal set of event sign up tests to cover:

- Sign up (opt out of mailing list)
- Sign up (opt into mailing list)
- Sign up and then perform a matchback and complete

The tests are setup so that if we run into the situation when there are no
Train to Teach events on the website the tests will still pass. To avoid false
positives we ensure the selector we are using to target an event from the
listings is still accurate (i.e. returns non-TTT events at least).