Support new matchback logic
Previously the API required the email/first/last to all match an existing
candidate, but it will now perform a fallback to just matching on the email
address. This means for new sign ups we need to use a unique email address and
for matching back we can simplify it by ensuring at least the email is the
same.