Update tests to reflect recent changes
There have been a couple of minor copy changes to the mailing list
and event sign up flows.

It also looks like the 'book a callback' button is now missing in
the preprod environment; I'm not sure if thats intentional with
the new welcome guide journey updates or not, but to get the
pipeline unblocked we can visit the book a callback wizard
directly.